create database in MaruAppFactory so it that will available to the rpc method factory
